Handle exception is org name already exists
Previously this would return a 500 error, as the 400 exception was
not handled from the API [1]. Note that:

- We tend to rely on exception messages to identify the error that
occurred [2][3], with services being a notable deviation [4]. I've
used the exception message approach, as this is more granular and
broadly consistent with the rest of the app.

- There is already code to cover this scenario when a user changes
the name of an existing organisation or service, but the mechanism
is different [5][6]. It makes sense to just get any error from the
call to try and create the organisation.

- The API mock is based on one for services [7], but I've chosen to
have it inline with the test, since we're unlikely to reuse it, and
it's clearer to have the test setup as part of the test.

[1]: https://github.com/bitzesty/notifications-api/blob/242d2f67ff3a3febbc41cdf6da1e18e14512c0ce/app/organisation/rest.py#L34-L47
[2]: https://github.com/bitzesty/notifications-admin/blob/0237c9b45bf72b91a75580bbffdafaa914ef5d8b/app/main/views/manage_users.py#L166
[3]: https://github.com/bitzesty/notifications-admin/blob/0237c9b45bf72b91a75580bbffdafaa914ef5d8b/app/main/views/templates.py#L499
[4]: https://github.com/bitzesty/notifications-admin/blob/0237c9b45bf72b91a75580bbffdafaa914ef5d8b/app/main/views/add_service.py#L31
[5]: https://github.com/bitzesty/notifications-admin/blob/0237c9b45bf72b91a75580bbffdafaa914ef5d8b/app/main/views/service_settings.py#L100-L102
[6]: https://github.com/bitzesty/notifications-admin/blob/0237c9b45bf72b91a75580bbffdafaa914ef5d8b/app/main/views/organisations.py#L263-L265
[7]:https://github.com/bitzesty/notifications-admin/blob/0237c9b45bf72b91a75580bbffdafaa914ef5d8b/tests/conftest.py#L572-L588